# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "description": "this is paro api.",
    "version": "1.0.0",
    "title": "Swagger parto"
  },
  "servers": [
    {
      "url": "https://api.parto.app"
    },
    {
      "url": "http://localhost:2216"
    },
    {
      "url": "https://dev.parto.app"
    }
  ],
...
  "paths": {
...
    "/profile/getProfile/{userId}/{lang}": {
      "get": {
        "tags": [
          "profile"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "description": "user id"
                }
              }
            }
          },
          {
            "in": "path",
            "name": "lang",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "language"
                }
              }
            }
          },
          {
            "name": "x-auth-token",
            "in": "header",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "an authorization header"
                }
              }
            }
          }
        ],
        "summary": "send user peofile information to user",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "description": "user profile information",
                      "properties": {
                        "user_id": {
                          "type": "integer"
                        },
                        "birthdate": {
                          "type": "string"
                        },
                        "avg_cycle_length": {
                          "type": "integer"
                        },
                        "avg_period_length": {
                          "type": "integer"
                        },
                        "avg_sleeping_hour": {
                          "type": "integer"
                        },
                        "pms_length": {
                          "type": "integer"
                        },
                        "height": {
                          "type": "integer"
                        },
                        "weight": {
                          "type": "integer"
                        },
                        "pregnant": {
                          "type": "integer"
                        },
                        "pregnancy_try": {
                          "type": "integer"
                        },
                        "locked": {
                          "type": "integer"
                        },
                        "last_period_date": {
                          "type": "string"
                        },
                        "blood_type": {
                          "type": "string"
                        },
                        "ovulation_prediction": {
                          "type": "integer"
                        },
                        "period_prediction": {
                          "type": "integer"
                        },
                        "red_days": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "user not found"
          }
        }
      }
    },
    "/profile/getPeriodInfo/{userId}/{lang}": {
      "get": {
        "tags": [
          "profile"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "description": "user id"
                }
              }
            }
          },
          {
            "in": "path",
            "name": "lang",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "language"
                }
              }
            }
          },
          {
            "name": "x-auth-token",
            "in": "header",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "an authorization header"
                }
              }
            }
          }
        ],
        "summary": "send user peofile information to user",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "description": "user profile information",
                      "properties": {
                        "user_id": {
                          "type": "integer"
                        },
                        "avg_cycle_length": {
                          "type": "integer"
                        },
                        "avg_period_length": {
                          "type": "integer"
                        },
                        "pms_length": {
                          "type": "integer"
                        },
                        "pregnant": {
                          "type": "integer"
                        },
                        "pregnancy_try": {
                          "type": "integer"
                        },
                        "last_period_date": {
                          "type": "string"
                        },
                        "ovulation_prediction": {
                          "type": "integer"
                        },
                        "period_prediction": {
                          "type": "integer"
                        },
                        "red_days": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "user not found"
          }
        }
      }
    },
    "/profile/getGeneralInfo/{userId}/{lang}": {
      "get": {
        "tags": [
          "profile"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "description": "user id"
                }
              }
            }
          },
          {
            "in": "path",
            "name": "lang",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "language"
                }
              }
            }
          },
          {
            "name": "x-auth-token",
            "in": "header",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "an authorization header"
                }
              }
            }
          }
        ],
        "summary": "send user peofile information to user",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "description": "user profile information",
                      "properties": {
                        "user_id": {
                          "type": "integer"
                        },
                        "birthdate": {
                          "type": "string"
                        },
                        "avg_sleeping_hour": {
                          "type": "integer"
                        },
                        "height": {
                          "type": "integer"
                        },
                        "weight": {
                          "type": "integer"
                        },
                        "locked": {
                          "type": "integer"
                        },
                        "blood_type": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "user not found"
          }
        }
      }
    },
    "/profile/editProfile/{userId}/{lang}": {
      "put": {
        "tags": [
          "profile"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "description": "user id"
                }
              }
            }
          },
          {
            "in": "path",
            "name": "lang",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "language"
                }
              }
            }
          },
          {
            "name": "x-auth-token",
            "in": "header",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "an authorization header"
                }
              }
            }
          }
        ],
        "summary": "edit user peofile information to user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "birthdate": {
                    "type": "string"
                  },
                  "cycleLength": {
                    "type": "integer"
                  },
                  "periodLength": {
                    "type": "integer"
                  },
                  "sleepingHour": {
                    "type": "integer"
                  },
                  "pmsLength": {
                    "type": "integer"
                  },
                  "height": {
                    "type": "integer"
                  },
                  "weight": {
                    "type": "integer"
                  },
                  "pregnant": {
                    "type": "integer"
                  },
                  "pregnancyTry": {
                    "type": "integer"
                  },
                  "isLock": {
                    "type": "integer"
                  },
                  "lastPeriodDate": {
                    "type": "string"
                  },
                  "bloodType": {
                    "type": "string"
                  },
                  "ovulationPred": {
                    "type": "integer"
                  },
                  "periodPred": {
                    "type": "integer"
                  },
                  "redDays": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "information not found"
          }
        }
      }
    },
    "/profile/editGeneralInfo/{userId}/{lang}": {
      "put": {
        "tags": [
          "profile"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "description": "user id"
                }
              }
            }
          },
          {
            "in": "path",
            "name": "lang",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "language"
                }
              }
            }
          },
          {
            "name": "x-auth-token",
            "in": "header",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "an authorization header"
                }
              }
            }
          }
        ],
        "summary": "edit user peofile information to user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "birthdate": {
                    "type": "string"
                  },
                  "sleepingHour": {
                    "type": "integer"
                  },
                  "height": {
                    "type": "integer"
                  },
                  "weight": {
                    "type": "integer"
                  },
                  "isLock": {
                    "type": "integer"
                  },
                  "bloodType": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "information not found"
          }
        }
      }
    },
    "/profile/editPeriodInfo/{userId}/{lang}": {
      "put": {
        "tags": [
          "profile"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "description": "user id"
                }
              }
            }
          },
          {
            "in": "path",
            "name": "lang",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "language"
                }
              }
            }
          },
          {
            "name": "x-auth-token",
            "in": "header",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "an authorization header"
                }
              }
            }
          }
        ],
        "summary": "edit user peofile information to user",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cycleLength": {
                    "type": "integer"
                  },
                  "periodLength": {
                    "type": "integer"
                  },
                  "pmsLength": {
                    "type": "integer"
                  },
                  "pregnant": {
                    "type": "integer"
                  },
                  "pregnancyTry": {
                    "type": "integer"
                  },
                  "lastPeriodDate": {
                    "type": "string"
                  },
                  "ovulationPred": {
                    "type": "integer"
                  },
                  "periodPred": {
                    "type": "integer"
                  },
                  "redDays": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "information not found"
          }
        }
      }
    },
    "/profile/deleteProfile/{userId}/{lang}": {
      "delete": {
        "tags": [
          "profile"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "description": "user id"
                }
              }
            }
          },
          {
            "in": "path",
            "name": "lang",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "language"
                }
              }
            }
          },
          {
            "name": "x-auth-token",
            "in": "header",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "an authorization header"
                }
              }
            }
          }
        ],
        "summary": "edit user peofile information to user",
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "information not found"
          }
        }
      }
    },
    "/profile/pregnancyMode/{userId}/{lang}": {
      "get": {
        "tags": [
          "profile"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "description": "user id"
                }
              }
            }
          },
          {
            "in": "path",
            "name": "lang",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "language"
                }
              }
            }
          },
          {
            "name": "x-auth-token",
            "in": "header",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "an authorization header"
                }
              }
            }
          }
        ],
        "summary": "send user peofile information to user",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "description": "user profile information",
                      "properties": {
                        "pregnant": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "user not found"
          }
        }
      }
    },
    "/profile/getUserStatus/{userId}/{lang}": {
      "get": {
        "tags": [
          "profile"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "description": "user id"
                }
              }
            }
          },
          {
            "in": "path",
            "name": "lang",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "language"
                }
              }
            }
          },
          {
            "name": "x-auth-token",
            "in": "header",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "an authorization header"
                }
              }
            }
          }
        ],
        "summary": "send user peofile information to user",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "description": "user profile information",
                      "properties": {
                        "pregnant": {
                          "type": "integer"
                        },
                        "pregnancy_try": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "user not found"
          }
        }
      }
    },
    "/profile/lockStatus/{userId}/{lang}": {
      "get": {
        "tags": [
          "profile"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "integer",
                  "description": "user id"
                }
              }
            }
          },
          {
            "in": "path",
            "name": "lang",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "language"
                }
              }
            }
          },
          {
            "name": "x-auth-token",
            "in": "header",
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "string",
                  "description": "an authorization header"
                }
              }
            }
          }
        ],
        "summary": "send user peofile information to user",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "description": "use lock info",
                      "properties": {
                        "locked": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "user not found"
          }
        }
      }
    },
...
  }
}